fixed button and sponsers
